+++ micronaut-maven-integration-tests/src/it/jsonschema-generate-url/invoker.properties
@@ -1,0 +1,1 @@
+invoker.goals = -ntp compile
