+++ pom.xml
@@ -78,7 +78,7 @@
         <micronaut.aot.version>2.6.0</micronaut.aot.version>
         <micronaut.test.resources.version>2.7.2</micronaut.test.resources.version>
         <micronaut.openapi.version>6.13.2</micronaut.openapi.version>
-        <micronaut.jsonschema.version>1.3.1-SNAPSHOT</micronaut.jsonschema.version>
+        <micronaut.jsonschema.version>1.4.0-SNAPSHOT</micronaut.jsonschema.version>
 
         <maven.resources-plugin.version>3.3.1</maven.resources-plugin.version>
         <maven.archiver-plugin.version>3.6.0</maven.archiver-plugin.version>
